rename: 폴더 컨벤션을 파스칼 케이스에서 케밥 케이스로 변경합니다.
