+++ src/routes/index.ts
@@ -1,1 +1,1 @@
-export * from './Routes';
+export * from './routes';
